Add `slack-archive.yaml` in the root as a fix to #1
As @elonen has kindly pointed out, if one has not access to a markdown
interpreter, nor to the GitHub page for the repository, the yaml snippet
in the README could be copied with unwanted additional whitespace, due
to markdown source syntax for vertical alignment. For sure we don't want
the users to disentangle these irrelevant technicalities about how .md
files embed code snippets, so we provide also a single .yaml file, which
could be easily uploaded to slack website or copied all in once, without
worrying about whitespace.

This way every user should have a straightforward way to retrieve the
yaml lines:

- the webpage adept would just click on the rendered button, to copy the
  correct indented code automagically

- the local clone adept would just use the provided slack-archive.yaml

Thanks again @elonen!

------------------------------------------------------------------------
This solves and closes (without merge) pull request #1
------------------------------------------------------------------------